+++ backend/spring-routie/src/main/java/routie/logging/config/DevLoggingConfig.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import routie.logging.aspect.RequestLoggingAspect;
+import routie.logging.infrastructure.aspect.RequestLoggingAspect;
 import routie.logging.domain.LogDataBuilder;
-import routie.logging.logger.ClientRequestLogger;
+import routie.logging.service.ClientRequestLogger;
 
 @Configuration
 @Profile("dev")
